remove `beardedspice` cask; see below for deets
there is an annoying bug (feature?!) where if beardedspice is active & spotify
is open on my desktop, when my laptop goes to sleep (or, to be more precise,
within ~1.5 minute window after laptop sleeps) the Spotify content is paused.
this was the source of the "phantom" pauses that I was experiencing with Spotify,
even when I was playing it on a different device (e.g. phone). so I'm removing
this cask. IIRC I initially installed this in order to force stop the play/pause
button on my macbook from opening the iTunes/Music app; I remember this being
the cleanest solution from my research back then. but interestingly even after
uninstalling this cask, I still no longer see the button-opens-app behavior, so
I wonder why this is the case. anyways, for now, just going to remove to stop
stupid bug described above, & can reconsider reinstalling if undesired keyboard
behavior comes back.