add confirmation step to upgrade & install steps for either casks or packages
got bit in the ass when tmux got upgraded to 3.6 while using it, so figured it'd be worth adding these confirmation steps when running this script so that I confirm that the upgrades look fine & I have a change to possibly look at change logs if needed before continuing. potential next step here would be to be more granular in this confirmation step, e.g. "yes continue upgrading all packages, except for these ..."